back/requirements.txt
# Core Django
Django>=5.0,<5.1
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
django-cors-headers>=4.3.1

# Image Processing
Pillow>=10.2.0
django-cleanup>=8.0.0

# Authentication
djangorestframework-simplejwt>=5.3.1
djoser>=2.2.2

# WebSocket Support
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Additional Functionality
django-filter>=23.5
django-environ>=0.11.2

# Background Tasks
celery>=5.3.4
redis>=5.0.1

# Export to Excel, Word, PDF
openpyxl>=3.1.2
xlsxwriter>=3.1.9
python-docx>=1.1.0
reportlab>=4.0.9
weasyprint>=60.2

# Development Tools
django-debug-toolbar>=4.2.0
pytest>=7.4.3
pytest-django>=4.7.0
black>=23.12.1
flake8>=7.0.0


# Production Server
gunicorn>=21.2.0

# Monitoring (Optional)
sentry-sdk>=1.39.2